requirements.txt
duckduckgo-search>=6.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
networkx>=3.0
matplotlib>=3.7.0
pymupdf>=1.24.0
Pillow>=10.0.0
